backend/requirements.txt
opencv-python
numpy
ultralytics
openai>=1.0.0
python-dotenv
flask
flask-socketio
flask-cors
flask-talisman
cryptography